docs: complete rustdoc documentation with runnable examples and unit annotations
Enhance crate documentation for publish readiness:

lib.rs:
- Add runnable quick-start example showing full workflow: token setup,
  config creation, factory pool construction, and swap execution
- Add architecture diagram (text art) showing layer relationships
- Add module guide table linking to all public modules

factory/mod.rs → factory/default_factory.rs:
- Extract DefaultPoolFactory struct, impl, and tests into dedicated file
- mod.rs now contains only module-level docs, module declaration, and
  re-export (no logic in mod.rs)
- Convert module-level text example to runnable doc-test
- Convert DefaultPoolFactory struct example to runnable doc-test
- Add runnable doc example to DefaultPoolFactory::create method

pools/pool_box.rs:
- Convert PoolBox text example to runnable doc-test

pools/constant_product.rs:
- Add runnable doc example showing construction via FromConfig and swap

All pool structs — unit annotations added to state field docs:
- ConstantProductPool: reserve units, liquidity units, fee counter units
- ClmmPool: tick (dimensionless), sqrt_price (ratio), liquidity units,
  fee growth (scaled by 2^64)
- HybridPool: token units, amplification (1-10000), liquidity units,
  invariant D units
- WeightedPool: weights (basis points, sum=10000), balance units,
  liquidity units, fee counter units
- DynamicPool: token units, oracle price (quote/base ratio),
  k (dimensionless), liquidity units
- OrderBookPool: fee (basis points), tick_size (price units),
  lot_size (token units)

Doc-test count: 24 → 30 (+6 new runnable examples)
All text block examples converted to runnable doc-tests where practical.

Closes #37
